+++ features/groups/src/main/kotlin/es/pedrazamiguez/splittrip/features/group/presentation/component/SettlementDisputeBanner.kt
@@ -1,0 +1,41 @@
+package es.pedrazamiguez.splittrip.features.group.presentation.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
+import es.pedrazamiguez.splittrip.core.designsystem.icon.TablerIcons
+import es.pedrazamiguez.splittrip.core.designsystem.icon.outline.AlertTriangle
+import es.pedrazamiguez.splittrip.core.designsystem.presentation.component.text.BodyText
+
+@Composable
+internal fun SettlementDisputeBanner(
+    reason: String?,
+    modifier: Modifier = Modifier
+) {
+    if (reason == null) return
+
+    HorizontalDivider()
+    Row(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(MaterialTheme.spacing.Small),
+        verticalAlignment = Alignment.Top
+    ) {
+        Icon(
+            TablerIcons.Outline.AlertTriangle,
+            contentDescription = null,
+            tint = MaterialTheme.colorScheme.error,
+            modifier = Modifier.padding(top = MaterialTheme.spacing.ExtraSmall)
+        )
+        BodyText(
+            text = reason,
+            color = MaterialTheme.colorScheme.error
+        )
+    }
+}
