+++ features/groups/src/main/kotlin/es/pedrazamiguez/splittrip/features/group/presentation/component/SettlementEmptyText.kt
@@ -1,0 +1,30 @@
+package es.pedrazamiguez.splittrip.features.group.presentation.component
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import es.pedrazamiguez.splittrip.core.designsystem.foundation.spacing
+import es.pedrazamiguez.splittrip.features.group.R
+
+@Composable
+internal fun SettlementEmptyText(
+    show: Boolean,
+    modifier: Modifier = Modifier
+) {
+    if (show) {
+        Spacer(modifier = Modifier.height(MaterialTheme.spacing.Large))
+        Text(
+            text = stringResource(R.string.settlement_overview_empty),
+            style = MaterialTheme.typography.bodyLarge,
+            textAlign = TextAlign.Center,
+            color = MaterialTheme.colorScheme.onSurfaceVariant,
+            modifier = modifier.fillMaxWidth()
+        )
+    }
+}
