+++ features/groups/src/main/res/values-es-rAN/strings.xml
@@ -118,4 +118,29 @@
     <string name="group_leave_error_balance">Tu çardo debe çêh çero antê de podêh abandonâh er grupo</string>
     <string name="group_leave_error_admin">Como creadôh der grupo, no puedê abandonâl-lo. Elimínalo en çu lugâh.</string>
     <string name="group_leave_error_general">No çe pudo abandonâh er grupo. Inténtalo de nuebo.</string>
+    <string name="settlement_overview_title">Liquidaçionê</string>
+    <string name="settlement_overview_subtitle">Rebiça y reçuerbe lâ liquidaçionê pendientê antê de çerrâh er biahe</string>
+    <string name="settlement_overview_section_pending">Pendientê</string>
+    <string name="settlement_overview_section_disputed">Dîpputâh</string>
+    <string name="settlement_overview_section_resolved">Reçuertâ</string>
+    <string name="settlement_overview_empty">Toâ lâ liquidaçionê reçuertâ</string>
+    <string name="settlement_overview_confirm">Confirmâh</string>
+    <string name="settlement_overview_dispute">Dîpputâh</string>
+    <string name="settlement_overview_dispute_title">Dîpputâh liquidaçión</string>
+    <string name="settlement_overview_dispute_reason_hint">Dêccribe por qué dîpputâ êtta liquidaçión</string>
+    <string name="settlement_overview_dispute_submit">Embiâh dîpputa</string>
+    <string name="settlement_overview_dispute_cancel">Cançelâh</string>
+    <string name="settlement_overview_dispute_success">Liquidaçión dîpputá corrêttamente</string>
+    <string name="settlement_overview_confirm_success">Liquidaçión confirmá corrêttamente</string>
+    <string name="settlement_overview_close_trip">Çerrâh biahe</string>
+    <string name="settlement_overview_closing_trip">Çerrando biahe…</string>
+    <string name="settlement_overview_trip_closed">¡Biahe çerrao corrêttamente!</string>
+    <string name="settlement_overview_error_confirm">Errôh ar confirmâh la liquidaçión. Inténtalo de nuebo</string>
+    <string name="settlement_overview_error_dispute">Errôh ar dîpputâh la liquidaçión. Inténtalo de nuebo</string>
+    <string name="settlement_overview_error_archive">Errôh ar çerrâh er biahe. Inténtalo de nuebo</string>
+    <string name="settlement_overview_error_loading">No pudimô cargâh lô datô de liquidaçión. Inténtalo de nuebo</string>
+    <string name="settlement_overview_status_suggested">Pendiente</string>
+    <string name="settlement_overview_status_confirmed_by_payer">Êpperando cobradôh</string>
+    <string name="settlement_overview_status_disputed">Dîpputá</string>
+    <string name="settlement_overview_status_resolved">Reçuerta</string>
 </resources>
