+++ features/groups/src/main/res/values-es/strings.xml
@@ -143,4 +143,31 @@
     <string name="group_leave_error_balance">Tu saldo debe ser cero antes de poder abandonar el grupo</string>
     <string name="group_leave_error_admin">Como creador del grupo, no puedes abandonarlo. Elimínalo en su lugar.</string>
     <string name="group_leave_error_general">No se pudo abandonar el grupo. Inténtalo de nuevo.</string>
+
+    <!-- Settlement Overview -->
+    <string name="settlement_overview_title">Liquidaciones</string>
+    <string name="settlement_overview_subtitle">Revisa y resuelve las liquidaciones pendientes antes de cerrar el viaje</string>
+    <string name="settlement_overview_section_pending">Pendientes</string>
+    <string name="settlement_overview_section_disputed">Disputadas</string>
+    <string name="settlement_overview_section_resolved">Resueltas</string>
+    <string name="settlement_overview_empty">Todas las liquidaciones resueltas</string>
+    <string name="settlement_overview_confirm">Confirmar</string>
+    <string name="settlement_overview_dispute">Disputar</string>
+    <string name="settlement_overview_dispute_title">Disputar liquidación</string>
+    <string name="settlement_overview_dispute_reason_hint">Describe por qué disputas esta liquidación</string>
+    <string name="settlement_overview_dispute_submit">Enviar disputa</string>
+    <string name="settlement_overview_dispute_cancel">Cancelar</string>
+    <string name="settlement_overview_dispute_success">Liquidación disputada correctamente</string>
+    <string name="settlement_overview_confirm_success">Liquidación confirmada correctamente</string>
+    <string name="settlement_overview_close_trip">Cerrar viaje</string>
+    <string name="settlement_overview_closing_trip">Cerrando viaje…</string>
+    <string name="settlement_overview_trip_closed">¡Viaje cerrado correctamente!</string>
+    <string name="settlement_overview_error_confirm">Error al confirmar la liquidación. Inténtalo de nuevo</string>
+    <string name="settlement_overview_error_dispute">Error al disputar la liquidación. Inténtalo de nuevo</string>
+    <string name="settlement_overview_error_archive">Error al cerrar el viaje. Inténtalo de nuevo</string>
+    <string name="settlement_overview_error_loading">No pudimos cargar los datos de liquidación. Inténtalo de nuevo</string>
+    <string name="settlement_overview_status_suggested">Pendiente</string>
+    <string name="settlement_overview_status_confirmed_by_payer">Esperando cobrador</string>
+    <string name="settlement_overview_status_disputed">Disputada</string>
+    <string name="settlement_overview_status_resolved">Resuelta</string>
 </resources>
